Added lnurl feature flag

+++ rust/src/api/btc_ln.rs
@@ -180,7 +180,7 @@ impl BtcLnSwap {
         let create_reverse_req = if out_address.is_some() {
             let address = out_address.unwrap();
             boltz_client::swaps::boltz::CreateReverseRequest {
-                invoice_amount: out_amount as u32,
+                invoice_amount: out_amount as u64,
                 from: "BTC".to_string(),
                 to: "BTC".to_string(),
                 preimage_hash: preimage.sha256,
